feat

+++ src/routers/file.ts
@@ -7,7 +7,7 @@ import { fileLogger } from '../common/logger'
 import { FileInfomation } from '../types'
 import Busboy from 'busboy'
 import db from '../common/database'
-import { v4 as uuidv4 } from 'uuid';
+import { v4 as uuidv4 } from 'uuid'
 
 /**
  * 获取文件夹信息
@@ -217,8 +217,8 @@ export async function encryptFile(req: Request, res: Response) {
 
 /**
  * add backup to database
- * @param req 
- * @param res 
+ * @param req
+ * @param res
  */
 export async function createBackup(req: Request, res: Response) {
   // check if the filePath parameter is missing
@@ -227,17 +227,17 @@ export async function createBackup(req: Request, res: Response) {
     res.send(errorInfo)
     return
   }
-  const backupName:string = uuidv4();
-  db.createBackup(filePath, backupName);
+  const backupName: string = uuidv4()
+  db.createBackup(filePath, backupName)
   // 复制文件到备份文件夹
-  fs.cpSync(filePath, path.join(config.backupPath, backupName), { recursive: true });
-  res.send('备份成功');
+  fs.cpSync(filePath, path.join(config.backupPath, backupName), { recursive: true })
+  res.send('备份成功')
 }
 
 /**
  * get backup list
- * @param req 
- * @param res 
+ * @param req
+ * @param res
  */
 export async function getBackupList(req: Request, res: Response) {
   const { exist, isDir, errorInfo, filePath } = checkFilePath(req)
@@ -245,17 +245,19 @@ export async function getBackupList(req: Request, res: Response) {
     res.send(errorInfo)
     return
   }
-  db.getBackupList(filePath).then((backupList: Array<string>) => {
-    res.send(backupList);
-  }).catch((err) => {
-    res.send(err);
-  })
+  db.getBackupList(filePath)
+    .then((backupList: Array<string>) => {
+      res.send(backupList)
+    })
+    .catch((err) => {
+      res.send(err)
+    })
 }
 
 /**
  * backup file
- * @param request 
- * @param response 
+ * @param request
+ * @param response
  */
 export async function backupFile(request: Request, response: Response) {
   const { exist, isDir, errorInfo, filePath } = checkFilePath(request)
@@ -263,18 +265,53 @@ export async function backupFile(request: Request, response: Response) {
     response.send(errorInfo)
     return
   }
-  const backupUUID = request.query.backupUUID.toString()
-  const backupFilePath = path.join(config.backupPath, backupUUID);
-  // 新建旧的备份
-  const backupName:string = uuidv4();
-  db.createBackup(filePath, backupName);
-  fs.cpSync(filePath, backupFilePath, { recursive: true });
+  const newUUID = request.query.backupUUID.toString()
+  const newPath = path.join(config.backupPath, newUUID)
+  const oldUUID: string = uuidv4()
+  const oldPath = path.join(config.backupPath, oldUUID)
+  db.createBackup(filePath, oldUUID)
+  fs.cpSync(filePath, oldPath, { recursive: true })
   // 删除需求备份
-  db.deleteBackup(backupUUID);
+  db.deleteBackup(newUUID)
   // 恢复需求备份
-  fs.rmSync(filePath, { recursive: true });
-  fs.cpSync(backupFilePath, filePath, { recursive: true });
-  fs.rmSync(backupFilePath, { recursive: true });
-  response.send('恢复成功');
+  fs.rmSync(filePath, { recursive: true })
+  fs.cpSync(newPath, filePath, { recursive: true })
+  fs.rmSync(newPath, { recursive: true })
+  response.send('恢复成功')
 }
 
+export async function deleteBackup(request: Request, response: Response) {
+  const { exist, isDir, errorInfo, filePath } = checkFilePath(request)
+  if (!exist || isDir) {
+    response.send(errorInfo)
+    return
+  }
+  const backupUUID = request.query.backupUUID.toString()
+  db.deleteBackup(backupUUID)
+  try {
+    fs.rmSync(path.join(config.backupPath, backupUUID), { recursive: true })
+  } catch (e) {
+    response.send('删除成功')
+    return
+  }
+  response.send('删除成功')
+}
+
+export async function deleteFile(request: Request, response: Response) {
+  const { exist, isDir, errorInfo, filePath } = checkFilePath(request)
+  if (!exist) {
+    response.send(errorInfo)
+    return
+  }
+  // check password
+  const password = request.query.password?.toString()
+  if ((await db.checkEncryptFile(filePath)) && !(await db.checkPassword(filePath, password))) {
+    response.send('password error')
+    return
+  }
+  // log
+  const ip = request.headers['x-forwarded-for'] || request.ip
+  fileLogger.info(`[web][${ip}] Delete file: ${request.query.filePath}`)
+  fs.rmSync(filePath, { recursive: true })
+  response.send('删除成功')
+}
